requirements.txt
discord.py==2.3.2
pytest==7.4.0
pygithub==2.2.0
